requirements.txt
fastapi==0.109.0
python-multipart==0.0.6
uvicorn==0.27.0
rembg==2.0.50
Pillow==10.2.0
python-dotenv==1.0.0
python-jose==3.3.0
cors==1.0.1